requirements.txt
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
numpy=1.19.2=pypi_0
numpy-base=1.20.2=py38hfae3a4d_0
numpydoc=1.1.0=pyhd3eb1b0_1
pandas=1.2.5=py38h295c915_0
pytorch=1.12.0=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
gpytorch=1.7.0=pyhd8ed1ab_1
scikit-image=0.18.1=py38ha9443f7_0
scikit-learn=1.0.1=py38h51133e4_0
scipy=1.4.1=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
astropy=4.3.1=py38h09021b7_0
matplotlib=3.4.3=py38h06a4308_0
matplotlib-base=3.4.3=py38hbbc1b5f_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
tqdm=4.62.3=pyhd3eb1b0_1
prettytable=2.1.0=pyhd8ed1ab_0
